requirements.txt
Whoosh~=2.7.4
requests~=2.32.3
aiohttp~=3.11.18
readability-lxml~=0.8.4.1
bs4~=0.0.2
beautifulsoup4~=4.13.4
uvicorn~=0.34.2
fastapi~=0.115.12
protobuf~=5.29.4
pydantic~=2.11.4
websockets